refactor(tests): add file extension on import paths
