+++ src/tests/CommandParserTest.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
-import { Teams } from '..';
-import { parser, BanchoResponseType } from '../parsers';
+import { Teams } from '../Player.js';
+import { parser, BanchoResponseType } from '../parsers/CommandParser.js';
 import log4js from 'log4js';
 
 describe("CommandParserTest", function () {
