+++ src/tests/LobbyKeeperTest.ts
@@ -1,10 +1,11 @@
 import { assert } from 'chai';
 import Sinon from 'sinon';
-import { Lobby, Roles } from '..';
-import { DummyIrcClient } from '../dummies';
-import { Mod, ScoreMode, TeamMode } from '../Modes';
-import { LobbyKeeper, LobbyKeeperOption, SlotKeeper } from '../plugins';
-import tu from './TestUtils';
+import { Lobby } from '../Lobby.js';
+import { Roles } from '../Player.js';
+import { DummyIrcClient } from '../dummies/DummyIrcClient.js';
+import { Mod, ScoreMode, TeamMode } from '../Modes.js';
+import { LobbyKeeper, LobbyKeeperOption, SlotKeeper } from '../plugins/LobbyKeeper.js';
+import tu from './TestUtils.js';
 
 describe("LobbyKeepserTest", function () {
