+++ src/tests/TestUtils.ts
@@ -1,7 +1,10 @@
-import { Lobby, logIrcEvent, Player, Roles } from '..';
-import { DummyIrcClient } from '../dummies';
-import { BanchoResponse, BanchoResponseType, MpSettingsResult } from '../parsers';
-import { TypedEvent } from '../libs';
+import { Lobby } from '../Lobby.js';
+import { logIrcEvent } from '../IIrcClient.js';
+import { Player, Roles } from '../Player.js';
+import { DummyIrcClient } from '../dummies/DummyIrcClient.js';
+import { BanchoResponse, BanchoResponseType } from '../parsers/CommandParser.js';
+import { MpSettingsResult } from '../parsers/MpSettingsParser.js';
+import { TypedEvent } from '../libs/TypedEvent.js';
 import { assert } from 'chai';
 import log4js from 'log4js';
